backend/requirements.txt
alembic==1.10.4
blinker==1.6.2
click==8.1.3
ecdsa==0.18.0
Flask==2.3.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
Werkzeug==2.3.2
